# fix(utils): improved parsing performance when using parseResults option