- **What kind of change does this PR introduce?** (Bug fix, feature, docs update, ...)
Increased parsing performance when using `parseResults` option.

* **What is the current behavior?** (You can also link to an open issue here)
The existing `utils/formatCallResults` is not optimal, and is especially slow for large datasets. A significant factor of this was that the function supported parsing results without a field mask, which causes it to recursively traverse all rows some of which can be very deeply nested. Other things we have to do include converting `google.protobuf.Value`'s to _normal values_ and removing undefined fields, not specified in the request. Multiply this by 100k rows and performance takes a significant hit.

- **What is the new behavior (if this is a feature change)?**
`formatCallResults` has been rewritten to do the majority of parsing in one pass of the entire dataset. It also now requires a field mask. Recursion is still required for parsing deeply nested objects/arrays, but it is now significantly faster. In the future, we should look at different ways of compiling the protocol buffers, as there may be lower-level ways to avoid this parsing - particularly for `google.protobuf.Value` fields e.g. converting `{clicks: { value: 2 }}` -> `{clicks: 2}`.

- **Bonus fix**
This enum issue has been resolved at the same time https://github.com/Opteo/google-ads-node/issues/43

* **Other information**:
These benchmarks don't include network time, which obviously is a factor for larger datasets. Some very large requests could take several seconds from Google's end. The data comprised of ads (ad_group_ad.ad), with many fields specified. `TET` stands for Total Execution Time, and all times were measured with Nodes `process.hrtime` and the `perf_hooks` module. This was recorded on a machine with 3 high performance vCPUs.

    | Rows | TET Before | TET After | Difference (%) |
    |------|------------|-----------|----------------|
    | 10k  | 1.632s     | 0.152s    | 165.9          |
    | 50k  | 8.376s     | 0.717s    | 168.5          |
    | 100k | 15.656s    | 1.101s    | 173.7          |
    | 250k | 41.768s    | 2.618s    | 176.4          |
    | 1m   | 168.337s   | 10.364s   | 176.8          |